+++ .storybook/stories/tagContextMenu/TagContextMenuWrapper.tsx
@@ -1,7 +1,8 @@
 import React, { useState } from 'react';
+
+import TagContextMenu from '../../../src/components/tagContextMenu/TagContextMenu';
+import TagIcon from '../../../src/elements/tagIcon/TagIcon';
 import { getIcon } from '../../../src/helpers/getIcon';
-import TagContextMenu from '../../../src/tagContextMenu/TagContextMenu';
-import TagIcon from '../../../src/tagIcon/TagIcon';
 
 export interface TagContextMenuWrapperProps {
     icon: string;
