+++ src/__tests__/ButtonWithPopover.test.tsx
@@ -1,8 +1,9 @@
-import { screen } from '@testing-library/dom';
 import { fireEvent, render } from '@testing-library/react';
-import React from 'react';
-import ButtonWithPopover from '../buttonWithPopover/ButtonWithPopover';
+
+import ButtonWithPopover from '../components/buttonWithPopover/ButtonWithPopover';
 import { DataInformation } from '../types/dataInformation';
+import React from 'react';
+import { screen } from '@testing-library/dom';
 
 test('should click show more and popover button should show open state', () => {
     const fetchDataToShow = jest.fn();
