+++ src/__tests__/DataInfoPopover.test.tsx
@@ -1,8 +1,9 @@
-import { screen, waitFor } from '@testing-library/dom';
 import { act, render } from '@testing-library/react';
-import React from 'react';
+import { screen, waitFor } from '@testing-library/dom';
+
+import DataInfoPopover from '../components/contextMenuPopover/DataInfoPopover';
 import { DataInformation } from '..';
-import DataInfoPopover from '../contextMenuPopover/DataInfoPopover';
+import React from 'react';
 
 beforeAll(() => {
     jest.useFakeTimers();
