+++ src/components/contextMenu/ContextMenu.tsx
@@ -1,6 +1,7 @@
 import React, { CSSProperties } from 'react';
+
 import TagContextMenu from '../tagContextMenu/TagContextMenu';
-import TagIcon from '../tagIcon/TagIcon';
+import TagIcon from '../../elements/tagIcon/TagIcon';
 import style from './contextMenu.module.css';
 
 interface ContextMenuProps {
