+++ src/elements/icon/Icon.tsx
@@ -1,5 +1,6 @@
-import { Icon as EdsIcon } from '@equinor/eds-core-react';
 import * as icons from '@equinor/eds-icons';
+
+import { Icon as EdsIcon } from '@equinor/eds-core-react';
 import React from 'react';
 
 export interface IconProps {
