+++ src/elements/tagIcon/tagIcon.module.css
